Refactor flight import & export plugins
