+++ src/PluginManager/CMakeLists.txt
@@ -14,19 +14,22 @@ target_sources(${LIBRARY_NAME}
         include/PluginManager/Csv.h
         include/PluginManager/SkyConnectIntf.h
         include/PluginManager/PluginIntf.h
-        include/PluginManager/FlightExportIntf.h
-        include/PluginManager/FlightImportIntf.h
+
         include/PluginManager/SkyConnectManager.h src/SkyConnectManager.cpp
         include/PluginManager/AbstractSkyConnect.h src/AbstractSkyConnect.cpp
         include/PluginManager/PluginManager.h src/PluginManager.cpp
         include/PluginManager/Export.h src/Export.cpp
         include/PluginManager/PluginBase.h src/PluginBase.cpp
-        include/PluginManager/FlightImportPluginBase.h src/FlightImportPluginBase.cpp
-        include/PluginManager/FlightImportPluginBaseSettings.h src/FlightImportPluginBaseSettings.cpp
-        include/PluginManager/BasicFlightImportDialog.h src/BasicFlightImportDialog.cpp src/BasicFlightImportDialog.ui
-        include/PluginManager/FlightExportPluginBase.h src/FlightExportPluginBase.cpp
-        include/PluginManager/FlightExportPluginBaseSettings.h src/FlightExportPluginBaseSettings.cpp
-        include/PluginManager/BasicFlightExportDialog.h src/BasicFlightExportDialog.cpp src/BasicFlightExportDialog.ui
+        # Flight import & export plugins
+        include/PluginManager/Flight/FlightExportIntf.h
+        include/PluginManager/Flight/FlightImportIntf.h
+        include/PluginManager/Flight/FlightImportPluginBase.h src/Flight/FlightImportPluginBase.cpp
+        include/PluginManager/Flight/FlightImportPluginBaseSettings.h src/Flight/FlightImportPluginBaseSettings.cpp
+        include/PluginManager/Flight/BasicFlightImportDialog.h src/Flight/BasicFlightImportDialog.cpp src/Flight/BasicFlightImportDialog.ui
+        include/PluginManager/Flight/FlightExportPluginBase.h src/Flight/FlightExportPluginBase.cpp
+        include/PluginManager/Flight/FlightExportPluginBaseSettings.h src/Flight/FlightExportPluginBaseSettings.cpp
+        include/PluginManager/Flight/BasicFlightExportDialog.h src/Flight/BasicFlightExportDialog.cpp src/Flight/BasicFlightExportDialog.ui
+        # Location import & export plugins
         include/PluginManager/LocationExportIntf.h
         include/PluginManager/LocationExportPluginBase.h src/LocationExportPluginBase.cpp
         include/PluginManager/LocationExportPluginBaseSettings.h src/LocationExportPluginBaseSettings.cpp
@@ -36,9 +39,9 @@ target_sources(${LIBRARY_NAME}
         include/PluginManager/LocationImportPluginBaseSettings.h src/LocationImportPluginBaseSettings.cpp
         include/PluginManager/BasicLocationImportDialog.h src/BasicLocationImportDialog.cpp src/BasicLocationImportDialog.ui
         # Modules
+        include/PluginManager/ModuleManager.h src/ModuleManager.cpp
         include/PluginManager/Module/ModuleIntf.h
-        include/PluginManager/Module/AbstractModule.h src/Module/AbstractModule.cpp
-        include/PluginManager/Module/ModuleManager.h src/Module/ModuleManager.cpp
+        include/PluginManager/Module/AbstractModule.h src/Module/AbstractModule.cpp        
         src/Module/DefaultModuleImpl.h src/Module/DefaultModuleImpl.cpp
         src/Module/DefaultModuleSettings.h src/Module/DefaultModuleSettings.cpp
         include/PluginManager/Module/ModuleBaseSettings.h src/Module/ModuleBaseSettings.cpp
