+++ src/PluginManager/include/PluginManager/Flight/FlightExportIntf.h
@@ -27,7 +27,7 @@
 
 #include <QtPlugin>
 
-#include "PluginIntf.h"
+#include "../PluginIntf.h"
 
 class Flight;
