+++ src/PluginManager/include/PluginManager/ModuleManager.h
@@ -43,7 +43,7 @@ class QLayout;
 
 #include <Kernel/Sort.h>
 #include <Kernel/QUuidHasher.h>
-#include "../PluginManagerLib.h"
+#include "PluginManagerLib.h"
 
 class ModuleIntf;
 struct ModuleManagerPrivate;
