+++ src/PluginManager/src/PluginManager.cpp
@@ -41,8 +41,8 @@
 #include <Model/Flight.h>
 #include <Persistence/Service/FlightService.h>
 #include <Persistence/Service/LocationService.h>
-#include "FlightImportIntf.h"
-#include "FlightExportIntf.h"
+#include <Flight/FlightImportIntf.h>
+#include <Flight/FlightExportIntf.h>
 #include "LocationImportIntf.h"
 #include "LocationExportIntf.h"
 #include "PluginManager.h"
