+++ src/Plugins/Flight/Export/CsvExport/src/CsvExportPlugin.h
@@ -36,8 +36,8 @@ class QString;
 class QDateTime;
 
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportIntf.h>
-#include <PluginManager/FlightExportPluginBase.h>
+#include <PluginManager/Flight/FlightExportIntf.h>
+#include <PluginManager/Flight/FlightExportPluginBase.h>
 
 struct FlightData;
 class Aircraft;
