+++ src/Plugins/Flight/Export/CsvExport/src/CsvExportSettings.cpp
@@ -29,7 +29,7 @@
 
 #include <Kernel/Enum.h>
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 #include "CsvExportSettings.h"
 
 namespace
