+++ src/Plugins/Flight/Export/IgcExport/src/IgcExportPlugin.h
@@ -37,8 +37,8 @@ class QString;
 class QDateTime;
 
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportIntf.h>
-#include <PluginManager/FlightExportPluginBase.h>
+#include <PluginManager/Flight/FlightExportIntf.h>
+#include <PluginManager/Flight/FlightExportPluginBase.h>
 
 struct FlightData;
 class Aircraft;
