+++ src/Plugins/Flight/Export/IgcExport/src/IgcExportSettings.h
@@ -31,7 +31,7 @@
 #include <QString>
 
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 
 struct IgcExportSettingsPrivate;
