+++ src/Plugins/Flight/Export/JsonExport/src/JsonExportSettings.cpp
@@ -32,7 +32,7 @@
 
 #include <Kernel/Enum.h>
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 #include "JsonExportSettings.h"
 
 struct JsonExportSettingsPrivate
