+++ src/Plugins/Flight/Export/KmlExport/src/KmlExportSettings.cpp
@@ -28,7 +28,7 @@
 
 #include <Kernel/Enum.h>
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 #include "KmlExportSettings.h"
 
 namespace
