+++ src/Plugins/Flight/Export/KmlExport/src/KmlExportSettings.h
@@ -31,7 +31,7 @@
 #include <QColor>
 
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 
 struct KmlExportSettingsPrivate;
