+++ src/Plugins/Flight/Export/SdLogExport/src/SdLogExportPlugin.h
@@ -36,8 +36,8 @@ class QIODevice;
 class QString;
 
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportIntf.h>
-#include <PluginManager/FlightExportPluginBase.h>
+#include <PluginManager/Flight/FlightExportIntf.h>
+#include <PluginManager/Flight/FlightExportPluginBase.h>
 
 class Aircraft;
 struct PositionData;
