+++ src/Plugins/Flight/Export/SdLogExport/src/SdLogExportSettings.cpp
@@ -29,7 +29,7 @@
 
 #include <Kernel/Enum.h>
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 #include "SdLogExportSettings.h"
 
 struct SdLogExportSettingsPrivate
