+++ src/Plugins/Flight/Export/SdLogExport/src/SdLogExportSettings.h
@@ -30,7 +30,7 @@
 #include <QObject>
 
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 
 struct SdLogExportSettingsPrivate;
