+++ src/Plugins/Location/Export/CsvLocationExport/src/CsvLocationExportSettings.cpp
@@ -32,7 +32,7 @@
 
 #include <Kernel/Enum.h>
 #include <Kernel/Settings.h>
-#include <PluginManager/FlightExportPluginBaseSettings.h>
+#include <PluginManager/Flight/FlightExportPluginBaseSettings.h>
 #include "CsvLocationExportSettings.h"
 
 namespace
